Repetitions done
I made a function with several loops to get different values depending on what i need. Starting off with a loop who makes all numbers on the page to an array, then a loop that creates an object where the key is the number and it's value its the repetition. Moving on to a loop that gets the highest value inside of the object and gives it to mostRepeated. Then I make another loop that creates a new object and so on.