Find same finished
Simple code with add event listener that first holds the number clicked in a seperate variable. Then it will go in to another loop where it checks if it is the same as any other. If so the counter will be affected, it will get affected as you click because there will always be at least one number. Depending on the counter it gives two ways of text (preference by me). Then a very easy code to restore the background colors and text.